requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
requests==2.32.3
yoomoney==0.1.0
aiocryptopay==0.4.6
